requirements.txt
langchain==0.3.9
langchain-community==0.3.8
langchain-ollama==0.2.0
ollama==0.4.2
streamlit==1.40.2
pdfplumber==0.11.4
fpdf==1.7.2
python-docx==1.1.0
pydantic==2.10.2
